requirements.txt
streamlit
Pillow
pytesseract
pandas
gspread
google-auth-oauthlib